requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0

# Transcription
faster-whisper>=1.0.0

# Speaker diarization (optional — needed for --diarize)
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
av>=12.0.0           # audio decoding (m4a/aac support)

# Notion sync (optional — needed if NOTION_TOKEN is set)
notion-client>=2.0.0

# Speaker embeddings
scipy>=1.11.0
scikit-learn>=1.3.0
numpy>=1.24.0
